Improved reporting of hardware platform.
Due to a change of the serial flash chip just before production,
the firmware used to incorrectly identify a production device as
a prototype.  The old flash chip's ID was used as one of the
criteria in determining hardware platform.
This patch replaces the expected ID with the JEDEC ID of the
flash chip actually installed in the production devices.