Non-streaming mode deprecated.
The mode where the entire generated JPEG is written to flash and then served from the flash is moved under #if AT25DFX_JPEG and declared deprecated.
The streaming mode where JPEG is generated on demand has been in use since early prototypes and works fine.
AT25DFX_MEM will be re-used for another purpose, and the flash-buffered JPEG mode will not be maintained.