Cosmetic improvements.
Add ;capture=camera to the 'Choose file' button to enable QR code capture from camera on supported devices.
It is hard to reliably detect whether the user agent will allow capture from camera or not, so the button's label does not change for now.
Do not display the 'Scan with camera' button on systems which do not support getUserMedia.
Show empty <img> instead of <video> in the box when page is loaded.